Shift handling of user commands to Parser class
