+++ data/oreo.txt
@@ -1,3 +1,5 @@
-D|0|return book|2002-11-17
+D|1|return book|2002-11-17
 T|1|sleep
-E|1|workshop|2025-08-30|2025-08-30
+E|0|workshop|2025-08-30|2025-08-30
+T|0|run
+D|0|ip|2025-08-29
